refactor: Move subjects into models
