+++ packages/neuron-wallet/src/startup/sync-block-task/task.ts
@@ -4,7 +4,7 @@ import initConnection from '../../database/chain/ormconfig'
 import { initConnection as initAddressConnection } from '../../database/address/ormconfig'
 import AddressService from '../../services/addresses'
 import LockUtils from '../../utils/lock-utils'
-import AddressesUsedSubject from '../../subjects/addresses-used-subject'
+import AddressesUsedSubject from '../../models/subjects/addresses-used-subject'
 import BlockListener from '../../services/sync/block-listener'
 import { NetworkWithID } from '../../services/networks'
